IGAPP-91: Upgrade react-navigation
