+++ native/package.json
@@ -72,9 +72,9 @@
     "react-native-static-server": "^0.5.0",
     "react-native-vector-icons": "^7.1.0",
     "react-native-webview": "^10.9.1",
-    "react-navigation": "^3.10.2",
+    "react-navigation": "^4.4.2",
     "react-navigation-header-buttons": "^3.0.5",
-    "react-navigation-stack": "^1.4.0",
+    "react-navigation-stack": "^1.7.3",
     "react-redux": "^7.1.3",
     "redux": "^4.0.5",
     "redux-actions": "^2.6.5",
