+++ native/src/modules/layout/components/Header.js
@@ -5,7 +5,7 @@ import { Share } from 'react-native'
 import styled from 'styled-components/native'
 import { type StyledComponent } from 'styled-components'
 import { Item } from 'react-navigation-header-buttons'
-import HeaderBackButton from 'react-navigation-stack/lib/module/views/Header/HeaderBackButton'
+import { HeaderBackButton } from 'react-navigation-stack'
 import type { NavigationDescriptor, NavigationScene, NavigationScreenProp } from 'react-navigation'
 import type { ThemeType } from '../../theme/constants'
 import type { TFunction } from 'react-i18next'
