+++ native/src/modules/layout/components/SettingsHeader.js
@@ -3,7 +3,7 @@
 import * as React from 'react'
 import styled from 'styled-components/native'
 import { type StyledComponent } from 'styled-components'
-import HeaderBackButton from 'react-navigation-stack/lib/module/views/Header/HeaderBackButton'
+import { HeaderBackButton } from 'react-navigation-stack'
 import type { ThemeType } from '../../theme/constants'
 import type { TFunction } from 'react-i18next'
 import type { NavigationScreenProp } from 'react-navigation'
