+++ native/src/modules/layout/components/TransparentHeader.js
@@ -6,7 +6,7 @@ import styled from 'styled-components/native'
 import { type StyledComponent } from 'styled-components'
 import type { NavigationScreenProp } from 'react-navigation'
 import type { ThemeType } from '../../theme/constants'
-import HeaderBackButton from 'react-navigation-stack/lib/module/views/Header/HeaderBackButton'
+import { HeaderBackButton } from 'react-navigation-stack'
 import { HeaderButtons, HeaderButton, Item } from 'react-navigation-header-buttons'
 import MaterialIcon from 'react-native-vector-icons/MaterialIcons'
 import type { TFunction } from 'react-i18next'
